backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
drf-extra-fields==3.7.0
pillow==11.2.1
django-bootstrap5==25.1
pytest==8.3.5
pytest-django==4.11.1
filetype==1.2.0
packaging==25.0
pluggy==1.6.0
gunicorn==23.0.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.9 